Packages reorganization for the notes screen
