+++ app/src/main/java/com/bhaktaprogram/notehelper/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.bhaktaprogram.notehelper.di.DaggerNotesComponent
+import com.bhaktaprogram.notehelper.features.notes.NotesRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers.Default
 import kotlinx.coroutines.launch
