+++ app/src/main/java/com/bhaktaprogram/notehelper/di/NotesComponent.kt
@@ -1,6 +1,10 @@
 package com.bhaktaprogram.notehelper.di
 
-import com.bhaktaprogram.notehelper.*
+import com.bhaktaprogram.notehelper.MainActivity
+import com.bhaktaprogram.notehelper.features.notes.NotesDataSource
+import com.bhaktaprogram.notehelper.features.notes.NotesDataSourceFake
+import com.bhaktaprogram.notehelper.features.notes.NotesRepository
+import com.bhaktaprogram.notehelper.features.notes.NotesRepositoryImpl
 import dagger.Binds
 import dagger.Component
 import dagger.Module
