+++ app/src/main/java/com/bhaktaprogram/notehelper/features/notes/NoteInfo.kt
@@ -1,4 +1,4 @@
-package com.bhaktaprogram.notehelper
+package com.bhaktaprogram.notehelper.features.notes
 
 data class NoteInfo(
     val id: Int,
