+++ app/src/main/java/com/bhaktaprogram/notehelper/features/notes/NotesDataSource.kt
@@ -1,4 +1,4 @@
-package com.bhaktaprogram.notehelper
+package com.bhaktaprogram.notehelper.features.notes
 
 interface NotesDataSource {
