+++ app/src/main/java/com/bhaktaprogram/notehelper/features/notes/NotesDataSourceFake.kt
@@ -1,4 +1,4 @@
-package com.bhaktaprogram.notehelper
+package com.bhaktaprogram.notehelper.features.notes
 
 import javax.inject.Inject
