+++ app/src/main/java/com/bhaktaprogram/notehelper/features/notes/NotesRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.bhaktaprogram.notehelper
+package com.bhaktaprogram.notehelper.features.notes
 
 import javax.inject.Inject
